requirements.txt
asgiref                       3.8.1
cffi                          1.17.1
cryptography                  45.0.4
Django                        5.2.3
djangorestframework           3.16.0
djangorestframework_simplejwt 5.5.0
dnspython                     2.7.0
drf-yasg                      1.21.10
inflection                    0.5.1
mongoengine                   0.29.1
packaging                     25.0
pip                           25.1.1
pycparser                     2.22
PyJWT                         2.9.0
pymongo                       4.13.2
python-dotenv                 1.1.0
pytz                          2025.2
PyYAML                        6.0.2
sqlparse                      0.5.3
tzdata                        2025.2
uritemplate                   4.2.0